test: extract provider sections to tests/providers/ (batch A, #1440)
