+++ test-all.mjs
@@ -2801,488 +2801,6 @@ try {
   fail(`follow-up cadence module crashed: ${e.message}`);
 }
 
-// ── 13. PROVIDERS — SmartRecruiters ─────────────────────────────────
-
-console.log('\n13. Provider — smartrecruiters');
-
-try {
-  const sr = (await import(pathToFileURL(join(ROOT, 'providers/smartrecruiters.mjs')).href)).default;
-  const { parseSmartRecruitersResponse } = await import(pathToFileURL(join(ROOT, 'providers/smartrecruiters.mjs')).href);
-
-  if (sr.id === 'smartrecruiters') pass('smartrecruiters.id is "smartrecruiters"');
-  else fail(`smartrecruiters.id is ${JSON.stringify(sr.id)}`);
-
-  const hitCareers = sr.detect({ name: 'Adyen', careers_url: 'https://careers.smartrecruiters.com/adyen' });
-  if (hitCareers && hitCareers.url.startsWith('https://api.smartrecruiters.com/v1/companies/adyen/postings')) {
-    pass('smartrecruiters.detect() resolves careers.smartrecruiters.com/<slug> → api URL');
-  } else {
-    fail(`smartrecruiters.detect(careers) returned ${JSON.stringify(hitCareers)}`);
-  }
-
-  const hitJobs = sr.detect({ name: 'X', careers_url: 'https://jobs.smartrecruiters.com/x' });
-  if (hitJobs && hitJobs.url.startsWith('https://api.smartrecruiters.com/v1/companies/x/postings')) {
-    pass('smartrecruiters.detect() also handles jobs.smartrecruiters.com');
-  } else {
-    fail(`smartrecruiters.detect(jobs) returned ${JSON.stringify(hitJobs)}`);
-  }
-
-  if (sr.detect({ name: 'X', careers_url: 'https://example.com/careers' }) === null) {
-    pass('smartrecruiters.detect() returns null for non-SR URLs');
-  } else {
-    fail('smartrecruiters.detect() should return null for non-SR URLs');
-  }
-
-  // entry.api precedence: a branded careers_url is kept while the SR slug is
-  // pinned via api: (mirrors greenhouse/ashby).
-  const hitApi = sr.detect({
-    name: 'Continental',
-    careers_url: 'https://jobs.continental.com',
-    api: 'https://careers.smartrecruiters.com/Continental',
-  });
-  if (hitApi && hitApi.url.startsWith('https://api.smartrecruiters.com/v1/companies/Continental/postings')) {
-    pass('smartrecruiters.detect() honors api: over a branded careers_url');
-  } else {
-    fail(`smartrecruiters.detect(api-pinned) returned ${JSON.stringify(hitApi)}`);
-  }
-
-  // parseSmartRecruitersResponse
-  const sample = {
-    content: [
-      {
-        id: 'abc-123',
-        name: 'Senior PM',
-        ref: 'https://api.smartrecruiters.com/v1/companies/sgs/postings/abc-123',
-        location: { fullLocation: 'Geneva, Switzerland', remote: false },
-      },
-      {
-        id: 'def-456',
-        name: 'Remote AI Engineer',
-        ref: 'https://api.smartrecruiters.com/v1/companies/sgs/postings/def-456',
-        location: { city: 'Paris', country: 'France', remote: true },
-      },
-      {
-        id: 'ghi-789',
-        name: 'No-ref Role',
-        location: { fullLocation: 'Berlin, Germany' },
-      },
-    ],
-  };
-  const jobs = parseSmartRecruitersResponse(sample, 'SGS');
-  if (jobs.length === 3) pass('parseSmartRecruitersResponse extracts 3 jobs');
-  else fail(`parseSmartRecruitersResponse returned ${jobs.length} jobs`);
-
-  if (jobs[0]?.location === 'Geneva, Switzerland' && jobs[0]?.title === 'Senior PM') {
-    pass('parseSmartRecruitersResponse uses fullLocation when present');
-  } else {
-    fail(`row 0 = ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[1]?.location === 'Paris, France, Remote') {
-    pass('parseSmartRecruitersResponse builds location from city/country/remote when no fullLocation');
-  } else {
-    fail(`row 1 location = ${JSON.stringify(jobs[1]?.location)}, expected "Paris, France, Remote"`);
-  }
-
-  if (jobs[0]?.url === 'https://jobs.smartrecruiters.com/sgs/postings/abc-123') {
-    pass('parseSmartRecruitersResponse rewrites api.smartrecruiters.com → jobs.smartrecruiters.com');
-  } else {
-    fail(`row 0 url = ${JSON.stringify(jobs[0]?.url)}`);
-  }
-
-  if (jobs[2]?.url && jobs[2].url.startsWith('https://jobs.smartrecruiters.com/sgs/ghi-789')) {
-    pass('parseSmartRecruitersResponse falls back to synthetic URL when ref is missing');
-  } else {
-    fail(`row 2 url = ${JSON.stringify(jobs[2]?.url)}`);
-  }
-
-  // Empty input safety
-  if (parseSmartRecruitersResponse({}, 'X').length === 0) pass('empty {} input → empty result');
-  else fail('empty {} input should yield empty result');
-
-  if (parseSmartRecruitersResponse({ content: 'not an array' }, 'X').length === 0) {
-    pass('non-array content → empty result (no crash)');
-  } else {
-    fail('non-array content should yield empty result');
-  }
-
-  // careers_url with non-string value → detect() returns null without crashing
-  if (sr.detect({ name: 'X', careers_url: { foo: 'bar' } }) === null) {
-    pass('smartrecruiters.detect() returns null for non-string careers_url (object)');
-  } else {
-    fail('smartrecruiters.detect() should treat non-string careers_url as missing');
-  }
-
-  // Fallback URL when both ref AND id are missing → empty string (not "undefined" in URL)
-  const noRefNoId = parseSmartRecruitersResponse(
-    { content: [{ name: 'Stranded Role' }] },
-    'X',
-  );
-  if (noRefNoId.length === 1 && noRefNoId[0].url === '') {
-    pass('parseSmartRecruitersResponse returns url="" when both ref and id are missing');
-  } else {
-    fail(`expected url='' when ref+id both missing, got ${JSON.stringify(noRefNoId[0])}`);
-  }
-
-  // SSRF: malicious URL with smartrecruiters hostname in the PATH (not host) must not be detected.
-  if (sr.detect({ name: 'Spoof', careers_url: 'https://evil.example/careers.smartrecruiters.com/slug' }) === null) {
-    pass('smartrecruiters.detect() rejects path-spoofed URLs');
-  } else {
-    fail('smartrecruiters.detect() must NOT misdetect path-spoofed URLs');
-  }
-
-  // SmartRecruiters: untrusted j.ref host falls through to fallback rather than rewriting
-  const bogusRef = parseSmartRecruitersResponse(
-    { content: [{ id: 'X1', name: 'Strange Role', ref: 'https://evil.example/v1/companies/x/postings/X1' }] },
-    'TestCo',
-  );
-  if (bogusRef[0]?.url && !bogusRef[0].url.includes('evil.example')) {
-    pass('parseSmartRecruitersResponse rejects untrusted j.ref host (falls through to fallback)');
-  } else {
-    fail(`untrusted j.ref leaked into url: ${JSON.stringify(bogusRef[0]?.url)}`);
-  }
-
-  // SmartRecruiters: companyName with spaces/symbols is slugified for the fallback URL
-  const slugifiedCompany = parseSmartRecruitersResponse(
-    { content: [{ id: 'X2', name: 'Strange Role' }] },
-    'My Acme & Co.',
-  );
-  if (slugifiedCompany[0]?.url === 'https://jobs.smartrecruiters.com/my-acme-co/X2-strange-role') {
-    pass('parseSmartRecruitersResponse slugifies the companyName for the fallback URL');
-  } else {
-    fail(`fallback URL not properly slugified: ${JSON.stringify(slugifiedCompany[0]?.url)}`);
-  }
-
-  // Pagination: fetch() loops until an empty page (or short page) is returned
-  let pageRequests = 0;
-  const pagedJobs = await sr.fetch(
-    { name: 'PagedCo', careers_url: 'https://careers.smartrecruiters.com/paged' },
-    {
-      transport: 'http',
-      fetchText: async () => { throw new Error('fetchText should not be called'); },
-      fetchJson: async (url) => {
-        pageRequests++;
-        const offset = parseInt(new URL(url).searchParams.get('offset') || '0', 10);
-        if (offset === 0) {
-          // Page 1: full page (100 items)
-          return { content: Array.from({ length: 100 }, (_, i) => ({ id: `P1-${i}`, name: `Role 1-${i}` })) };
-        }
-        if (offset === 100) {
-          // Page 2: short page (50 items) → loop stops after this
-          return { content: Array.from({ length: 50 }, (_, i) => ({ id: `P2-${i}`, name: `Role 2-${i}` })) };
-        }
-        // Should not be reached because page 2 was short
-        return { content: [] };
-      },
-    },
-  );
-  if (pageRequests === 2 && pagedJobs.length === 150) {
-    pass('smartrecruiters.fetch() paginates and aggregates results (2 pages → 150 total)');
-  } else {
-    fail(`pagination: pageRequests=${pageRequests}, total=${pagedJobs.length} (expected 2 requests / 150 results)`);
-  }
-
-  // Pagination stop condition: empty content terminates the loop
-  let emptyPageRequests = 0;
-  const emptyJobs = await sr.fetch(
-    { name: 'EmptyCo', careers_url: 'https://careers.smartrecruiters.com/empty' },
-    {
-      transport: 'http',
-      fetchText: async () => { throw new Error('fetchText should not be called'); },
-      fetchJson: async () => {
-        emptyPageRequests++;
-        return { content: [] };
-      },
-    },
-  );
-  if (emptyPageRequests === 1 && emptyJobs.length === 0) {
-    pass('smartrecruiters.fetch() stops on the first empty page');
-  } else {
-    fail(`empty pagination: requests=${emptyPageRequests}, total=${emptyJobs.length}`);
-  }
-
-} catch (e) {
-  fail(`smartrecruiters provider tests crashed: ${e.message}`);
-}
-
-// ── 14. PROVIDERS — Recruitee ───────────────────────────────────────
-
-console.log('\n14. Provider — recruitee');
-
-try {
-  const recruitee = (await import(pathToFileURL(join(ROOT, 'providers/recruitee.mjs')).href)).default;
-  const { parseRecruiteeResponse } = await import(pathToFileURL(join(ROOT, 'providers/recruitee.mjs')).href);
-
-  if (recruitee.id === 'recruitee') pass('recruitee.id is "recruitee"');
-  else fail(`recruitee.id is ${JSON.stringify(recruitee.id)}`);
-
-  const hit = recruitee.detect({ name: 'Channable', careers_url: 'https://channable.recruitee.com' });
-  if (hit && hit.url === 'https://channable.recruitee.com/api/offers/') {
-    pass('recruitee.detect() resolves <slug>.recruitee.com → api offers');
-  } else {
-    fail(`recruitee.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  if (recruitee.detect({ name: 'X', careers_url: 'https://example.com/careers' }) === null) {
-    pass('recruitee.detect() returns null for non-recruitee URLs');
-  } else {
-    fail('recruitee.detect() should return null for non-recruitee URLs');
-  }
-
-  // parseRecruiteeResponse
-  const sample = {
-    offers: [
-      { title: 'Senior PM', careers_url: 'https://channable.recruitee.com/o/senior-pm', city: 'Utrecht', country: 'Netherlands', remote: false },
-      { title: 'Backend Eng', url: 'https://channable.recruitee.com/o/backend', city: 'Amsterdam', country: 'Netherlands', remote: true },
-      { title: 'AI Lead', location: 'Remote, EMEA' },
-    ],
-  };
-  const jobs = parseRecruiteeResponse(sample, 'Channable');
-  if (jobs.length === 3) pass('parseRecruiteeResponse extracts 3 offers');
-  else fail(`parseRecruiteeResponse returned ${jobs.length} offers`);
-
-  if (jobs[0]?.title === 'Senior PM' && jobs[0]?.company === 'Channable' && jobs[0]?.url === 'https://channable.recruitee.com/o/senior-pm') {
-    pass('parseRecruiteeResponse prefers careers_url field over url');
-  } else {
-    fail(`row 0 = ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[1]?.location === 'Amsterdam, Netherlands, Remote') {
-    pass('parseRecruiteeResponse assembles city/country/remote when no location field');
-  } else {
-    fail(`row 1 location = ${JSON.stringify(jobs[1]?.location)}, expected "Amsterdam, Netherlands, Remote"`);
-  }
-
-  if (jobs[2]?.location === 'Remote, EMEA') {
-    pass('parseRecruiteeResponse uses explicit location field when present');
-  } else {
-    fail(`row 2 location = ${JSON.stringify(jobs[2]?.location)}`);
-  }
-
-  if (parseRecruiteeResponse({}, 'X').length === 0) pass('empty {} → empty result');
-  else fail('empty {} should yield empty result');
-
-  if (parseRecruiteeResponse({ offers: null }, 'X').length === 0) {
-    pass('null offers → empty result (no crash)');
-  } else {
-    fail('null offers should yield empty result');
-  }
-
-  // careers_url with non-string value → detect() returns null without crashing
-  if (recruitee.detect({ name: 'X', careers_url: null }) === null && recruitee.detect({ name: 'X', careers_url: 7 }) === null) {
-    pass('recruitee.detect() returns null for non-string careers_url (null and 7)');
-  } else {
-    fail('recruitee.detect() should treat non-string careers_url as missing');
-  }
-
-  // SSRF: malicious URL with recruitee.com in the PATH (not host) must not be detected.
-  if (recruitee.detect({ name: 'Spoof', careers_url: 'https://evil.example/channable.recruitee.com/foo' }) === null) {
-    pass('recruitee.detect() rejects path-spoofed URLs');
-  } else {
-    fail('recruitee.detect() must NOT misdetect path-spoofed URLs');
-  }
-
-  // Per-offer URL validation: custom-domain https URLs are KEPT (Recruitee
-  // tenants serve postings on their own domain, e.g. careers.hostaway.com);
-  // only non-https and malformed/missing URLs are dropped. The per-offer URL
-  // is display-only and not host-locked to *.recruitee.com — see #recruitee.
-  const offerUrlOffers = parseRecruiteeResponse(
-    {
-      offers: [
-        { title: 'Recruitee domain', careers_url: 'https://channable.recruitee.com/o/good' },
-        { title: 'Custom domain', careers_url: 'https://careers.hostaway.com/o/senior-backend' },
-        { title: 'Insecure', careers_url: 'http://channable.recruitee.com/o/insecure' },
-        { title: 'No URL field' },
-      ],
-    },
-    'Channable',
-  );
-  if (offerUrlOffers[0]?.url === 'https://channable.recruitee.com/o/good' && offerUrlOffers[1]?.url === 'https://careers.hostaway.com/o/senior-backend' && offerUrlOffers[2]?.url === '' && offerUrlOffers[3]?.url === '') {
-    pass('parseRecruiteeResponse keeps custom-domain https URLs, drops non-https and missing');
-  } else {
-    fail(`URL validation: row0=${JSON.stringify(offerUrlOffers[0]?.url)}, row1=${JSON.stringify(offerUrlOffers[1]?.url)}, row2=${JSON.stringify(offerUrlOffers[2]?.url)}, row3=${JSON.stringify(offerUrlOffers[3]?.url)}`);
-  }
-
-} catch (e) {
-  fail(`recruitee provider tests crashed: ${e.message}`);
-}
-
-// ── 14. PROVIDERS — Teamtailor ──────────────────────────────────────
-
-console.log('\n14. Provider — teamtailor');
-
-try {
-  const teamtailor = (await import(pathToFileURL(join(ROOT, 'providers/teamtailor.mjs')).href)).default;
-  const { parseTeamtailorFeed } = await import(pathToFileURL(join(ROOT, 'providers/teamtailor.mjs')).href);
-
-  if (teamtailor.id === 'teamtailor') pass('teamtailor.id is "teamtailor"');
-  else fail(`teamtailor.id is ${JSON.stringify(teamtailor.id)}`);
-
-  // detect() — auto-detection from a <slug>.teamtailor.com careers_url, with
-  // any path normalized to /jobs.rss.
-  const hit = teamtailor.detect({ name: 'Podimo', careers_url: 'https://podimo.teamtailor.com/jobs' });
-  if (hit && hit.url === 'https://podimo.teamtailor.com/jobs.rss') {
-    pass('teamtailor.detect() resolves <slug>.teamtailor.com → /jobs.rss feed');
-  } else {
-    fail(`teamtailor.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  if (teamtailor.detect({ name: 'X', careers_url: 'https://example.com/careers' }) === null) {
-    pass('teamtailor.detect() returns null for non-teamtailor URLs');
-  } else {
-    fail('teamtailor.detect() should return null for non-teamtailor URLs');
-  }
-
-  // non-string careers_url → detect() returns null without crashing
-  if (teamtailor.detect({ name: 'X', careers_url: null }) === null && teamtailor.detect({ name: 'X', careers_url: 7 }) === null) {
-    pass('teamtailor.detect() returns null for non-string careers_url (null and 7)');
-  } else {
-    fail('teamtailor.detect() should treat non-string careers_url as missing');
-  }
-
-  // SSRF: teamtailor.com in the PATH (not host) must not be detected.
-  if (teamtailor.detect({ name: 'Spoof', careers_url: 'https://evil.example/podimo.teamtailor.com/jobs' }) === null) {
-    pass('teamtailor.detect() rejects path-spoofed URLs');
-  } else {
-    fail('teamtailor.detect() must NOT misdetect path-spoofed URLs');
-  }
-
-  // non-https careers_url is rejected
-  if (teamtailor.detect({ name: 'X', careers_url: 'http://podimo.teamtailor.com/jobs' }) === null) {
-    pass('teamtailor.detect() rejects non-https careers_url');
-  } else {
-    fail('teamtailor.detect() should reject non-https careers_url');
-  }
-
-  // parseTeamtailorFeed — RSS with tt: locations block and branded job link
-  const sampleXml = [
-    '<?xml version="1.0" encoding="UTF-8"?>',
-    '<rss version="2.0" xmlns:tt="https://teamtailor.com/locations"><channel>',
-    '<title>Podimo</title>',
-    '<item>',
-    '  <title>Sales Director &amp; Lead</title>',
-    '  <link>https://careers.podimo.com/jobs/7950030-sales-director</link>',
-    '  <pubDate>Mon, 22 Jun 2026 13:45:57 +0200</pubDate>',
-    '  <remoteStatus>hybrid</remoteStatus>',
-    '  <tt:locations><tt:location><tt:city>Oslo</tt:city><tt:country>Norway</tt:country></tt:location></tt:locations>',
-    '</item>',
-    '<item>',
-    '  <title>Remote Engineer</title>',
-    '  <link>https://podimo.teamtailor.com/jobs/123-remote-engineer</link>',
-    '  <remoteStatus>fully</remoteStatus>',
-    '</item>',
-    '</channel></rss>',
-  ].join('\n');
-
-  const jobs = parseTeamtailorFeed(sampleXml, 'Podimo');
-  if (jobs.length === 2) pass('parseTeamtailorFeed extracts 2 jobs from 2-item feed');
-  else fail(`parseTeamtailorFeed returned ${jobs.length} jobs, expected 2`);
-
-  if (jobs[0]?.title === 'Sales Director & Lead' && jobs[0]?.company === 'Podimo' && jobs[0]?.url === 'https://careers.podimo.com/jobs/7950030-sales-director') {
-    pass('parseTeamtailorFeed decodes title entities, sets company, keeps branded-domain link');
-  } else {
-    fail(`row 0 = ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[0]?.location === 'Oslo, Norway') {
-    pass('parseTeamtailorFeed builds location from tt:city + tt:country');
-  } else {
-    fail(`row 0 location = ${JSON.stringify(jobs[0]?.location)}, expected "Oslo, Norway"`);
-  }
-
-  if (jobs[0]?.postedAt === Date.parse('Mon, 22 Jun 2026 13:45:57 +0200')) {
-    pass('parseTeamtailorFeed parses pubDate → postedAt epoch ms');
-  } else {
-    fail(`row 0 postedAt = ${JSON.stringify(jobs[0]?.postedAt)}`);
-  }
-
-  if (jobs[1]?.location === 'Remote' && jobs[1]?.postedAt === undefined) {
-    pass('parseTeamtailorFeed falls back to "Remote" for fully-remote item with no place/date');
-  } else {
-    fail(`row 1 = ${JSON.stringify(jobs[1])}`);
-  }
-
-  // Robustness
-  if (parseTeamtailorFeed('', 'X').length === 0) pass('empty input → empty result');
-  else fail('empty input should yield empty result');
-
-  if (parseTeamtailorFeed(null, 'X').length === 0) pass('null input → empty result (no crash)');
-  else fail('null input should yield empty result without crashing');
-
-  // A well-formed item with no <link> is skipped, not emitted with a blank URL.
-  const noLink = parseTeamtailorFeed('<item><title>Ghost</title></item>', 'X');
-  if (noLink.length === 0) pass('item without <link> is dropped');
-  else fail(`item without <link> should be dropped, got ${JSON.stringify(noLink)}`);
-
-  // fetch() pins the request to the teamtailor.com host on the happy path and
-  // must pass redirect:'error' (asserting the SSRF guard, not just the URL).
-  const fetchJobs = await teamtailor.fetch(
-    { name: 'Podimo', careers_url: 'https://podimo.teamtailor.com/jobs' },
-    {
-      transport: 'http',
-      fetchText: async (url, options) => {
-        if (url !== 'https://podimo.teamtailor.com/jobs.rss') {
-          throw new Error(`fetchText called with unexpected URL: ${url}`);
-        }
-        if (options?.redirect !== 'error') {
-          throw new Error(`fetchText called without redirect:'error': ${JSON.stringify(options)}`);
-        }
-        return sampleXml;
-      },
-      fetchJson: async () => { throw new Error('fetchJson should not be called'); },
-    },
-  );
-  if (fetchJobs.length === 2) pass('teamtailor.fetch() hits /jobs.rss with redirect:error and returns parsed jobs');
-  else fail(`teamtailor.fetch() returned ${fetchJobs.length} jobs, expected 2`);
-
-  // Branded careers domain: auto-detection must NOT claim it (stays pinned to
-  // *.teamtailor.com), but an explicit `provider: teamtailor` entry may fetch
-  // the same /jobs.rss off the branded host the user configured.
-  if (teamtailor.detect({ name: 'Podimo', careers_url: 'https://careers.podimo.com/jobs' }) === null) {
-    pass('teamtailor.detect() does NOT auto-claim a branded (non-teamtailor.com) host');
-  } else {
-    fail('teamtailor.detect() must not auto-detect branded hosts');
-  }
-
-  const brandedJobs = await teamtailor.fetch(
-    { name: 'Podimo', provider: 'teamtailor', careers_url: 'https://careers.podimo.com/jobs' },
-    {
-      transport: 'http',
-      fetchText: async (url, options) => {
-        if (url !== 'https://careers.podimo.com/jobs.rss') {
-          throw new Error(`fetchText called with unexpected URL: ${url}`);
-        }
-        if (options?.redirect !== 'error') {
-          throw new Error(`fetchText called without redirect:'error': ${JSON.stringify(options)}`);
-        }
-        return sampleXml;
-      },
-      fetchJson: async () => { throw new Error('fetchJson should not be called'); },
-    },
-  );
-  if (brandedJobs.length === 2) pass('explicit provider:teamtailor fetches /jobs.rss off a branded careers host');
-  else fail(`branded-host fetch returned ${brandedJobs.length} jobs, expected 2`);
-
-  // A branded host WITHOUT the explicit provider opt-in must still be refused by fetch().
-  let brandedRefused = false;
-  try {
-    await teamtailor.fetch(
-      { name: 'Podimo', careers_url: 'https://careers.podimo.com/jobs' },
-      {
-        transport: 'http',
-        fetchText: async () => { throw new Error('fetchText should not be reached'); },
-        fetchJson: async () => { throw new Error('fetchJson should not be called'); },
-      },
-    );
-  } catch {
-    brandedRefused = true;
-  }
-  if (brandedRefused) pass('teamtailor.fetch() refuses a branded host without explicit provider:teamtailor');
-  else fail('teamtailor.fetch() should refuse a branded host when not explicitly configured');
-
-} catch (e) {
-  fail(`teamtailor provider tests crashed: ${e.message}`);
-}
-
 // ── 14b. ADD-ENTRY (/career-ops add) ────────────────────────────────
 
 console.log('\n14b. add-entry.mjs (dedup + insertion)');
@@ -4808,254 +4326,6 @@ if (!/Claude can interact/i.test(applyModeText)) {
   fail('apply mode still uses Claude-specific wording');
 }
 
-// ── 15. PROVIDERS — SolidJobs ─────────────────────────────────────
-
-console.log('\n15. Provider — solidjobs');
-
-try {
-  const sj = (await import(pathToFileURL(join(ROOT, 'providers/solidjobs.mjs')).href)).default;
-
-  if (sj.id === 'solidjobs') pass('solidjobs.id is "solidjobs"');
-  else fail(`solidjobs.id is ${JSON.stringify(sj.id)}`);
-
-  // detect() matches valid SolidJobs API URL
-  const hit = sj.detect({ name: 'SJ', careers_url: 'https://solid.jobs/public-api/offers/it?campaign=career-ops' });
-  if (hit && hit.url) pass('solidjobs.detect() matches valid API URL');
-  else fail('solidjobs.detect() should match solid.jobs public-api URL');
-
-  // detect() rejects non-SolidJobs URL
-  if (sj.detect({ name: 'X', careers_url: 'https://example.com/jobs' }) === null) {
-    pass('solidjobs.detect() rejects non-SolidJobs URL');
-  } else {
-    fail('solidjobs.detect() must reject non-SolidJobs URLs');
-  }
-
-  // detect() rejects path-spoofed URL (solid.jobs in path, not hostname)
-  if (sj.detect({ name: 'X', careers_url: 'https://evil.example/solid.jobs/public-api/offers/it' }) === null) {
-    pass('solidjobs.detect() rejects path-spoofed URLs');
-  } else {
-    fail('solidjobs.detect() must NOT misdetect URLs with solid.jobs in the path');
-  }
-
-  // detect() returns null for non-string careers_url
-  if (sj.detect({ name: 'X', careers_url: 42 }) === null) {
-    pass('solidjobs.detect() returns null for non-string careers_url (42)');
-  } else {
-    fail('solidjobs.detect() should treat non-string careers_url as missing');
-  }
-
-  // detect() returns null for missing careers_url
-  if (sj.detect({ name: 'X' }) === null) {
-    pass('solidjobs.detect() returns null for missing careers_url');
-  } else {
-    fail('solidjobs.detect() should return null when careers_url is missing');
-  }
-
-  // fetch() parses { jobs: [...] } response with company from API
-  const fakeJobs = {
-    jobs: [
-      { title: 'Senior Dev', url: 'https://solid.jobs/o/abc123/career-ops', company: 'Acme Corp', locations: ['Warszawa', 'Remote'] },
-      { title: 'Junior Dev', url: 'https://solid.jobs/o/def456/career-ops', company: 'Beta Inc', locations: ['Kraków'] },
-    ],
-  };
-  const parsed = await sj.fetch(
-    { name: 'SolidJobs IT', careers_url: 'https://solid.jobs/public-api/offers/it?campaign=career-ops' },
-    { transport: 'http', fetchJson: async () => fakeJobs, fetchText: async () => '' },
-  );
-  if (parsed.length === 2) pass('solidjobs.fetch() returns 2 jobs from mock response');
-  else fail(`solidjobs.fetch() returned ${parsed.length} jobs, expected 2`);
-
-  if (parsed[0].company === 'Acme Corp') pass('solidjobs.fetch() uses j.company from API response');
-  else fail(`solidjobs.fetch() company is ${JSON.stringify(parsed[0].company)}, expected "Acme Corp"`);
-
-  if (parsed[0].location === 'Warszawa, Remote') pass('solidjobs.fetch() joins locations array');
-  else fail(`solidjobs.fetch() location is ${JSON.stringify(parsed[0].location)}, expected "Warszawa, Remote"`);
-
-  if (parsed[0].title === 'Senior Dev' && parsed[0].url === 'https://solid.jobs/o/abc123/career-ops') {
-    pass('solidjobs.fetch() maps title and url correctly');
-  } else {
-    fail(`solidjobs.fetch() title/url wrong: ${JSON.stringify(parsed[0])}`);
-  }
-
-  // fetch() falls back to entry.name when j.company is missing
-  const noCompanyJobs = { jobs: [{ title: 'Tester', url: 'https://solid.jobs/o/xyz/career-ops', locations: [] }] };
-  const fallback = await sj.fetch(
-    { name: 'SolidJobs IT', careers_url: 'https://solid.jobs/public-api/offers/it?campaign=career-ops' },
-    { transport: 'http', fetchJson: async () => noCompanyJobs, fetchText: async () => '' },
-  );
-  if (fallback[0].company === 'SolidJobs IT') pass('solidjobs.fetch() falls back to entry.name when j.company missing');
-  else fail(`solidjobs.fetch() fallback company is ${JSON.stringify(fallback[0].company)}`);
-
-  // fetch() handles empty locations array
-  if (fallback[0].location === '') pass('solidjobs.fetch() returns empty string for empty locations array');
-  else fail(`solidjobs.fetch() location for empty array is ${JSON.stringify(fallback[0].location)}`);
-
-  // fetch() rejects non-SolidJobs hostname (SSRF)
-  let ssrfRejected = false;
-  try {
-    await sj.fetch(
-      { name: 'Evil', careers_url: 'https://evil.com/public-api/offers/it' },
-      { transport: 'http', fetchJson: async () => { throw new Error('SSRF! should not reach here'); }, fetchText: async () => '' },
-    );
-  } catch (e) {
-    if (e.message.includes('untrusted hostname')) ssrfRejected = true;
-    else fail(`solidjobs.fetch() rejected with wrong error: ${e.message}`);
-  }
-  if (ssrfRejected) pass('solidjobs.fetch() rejects untrusted hostname (SSRF protection)');
-  else fail('solidjobs.fetch() should reject non-solid.jobs hostnames');
-
-  // fetch() throws on missing careers_url
-  let missingUrl = false;
-  try {
-    await sj.fetch(
-      { name: 'No URL' },
-      { transport: 'http', fetchJson: async () => ({}), fetchText: async () => '' },
-    );
-  } catch (e) {
-    if (e.message.includes('careers_url required')) missingUrl = true;
-    else fail(`solidjobs.fetch() missing URL error: ${e.message}`);
-  }
-  if (missingUrl) pass('solidjobs.fetch() throws on missing careers_url');
-  else fail('solidjobs.fetch() should throw when careers_url is missing');
-
-  // fetch() rejects HTTP (non-HTTPS) URL
-  let httpRejected = false;
-  try {
-    await sj.fetch(
-      { name: 'HTTP', careers_url: 'http://solid.jobs/public-api/offers/it' },
-      { transport: 'http', fetchJson: async () => { throw new Error('should not reach here'); }, fetchText: async () => '' },
-    );
-  } catch (e) {
-    if (e.message.includes('HTTPS')) httpRejected = true;
-    else fail(`solidjobs.fetch() HTTP rejection wrong error: ${e.message}`);
-  }
-  if (httpRejected) pass('solidjobs.fetch() rejects HTTP URLs (HTTPS enforcement)');
-  else fail('solidjobs.fetch() should reject non-HTTPS URLs');
-
-  // fetch() rejects malformed/unparseable URL
-  let malformedRejected = false;
-  try {
-    await sj.fetch(
-      { name: 'Bad', careers_url: 'not-a-url' },
-      { transport: 'http', fetchJson: async () => { throw new Error('should not reach here'); }, fetchText: async () => '' },
-    );
-  } catch (e) {
-    if (e.message.includes('invalid URL')) malformedRejected = true;
-    else fail(`solidjobs.fetch() malformed URL wrong error: ${e.message}`);
-  }
-  if (malformedRejected) pass('solidjobs.fetch() rejects malformed URLs');
-  else fail('solidjobs.fetch() should reject unparseable URLs');
-
-  // fetch() throws on unexpected API response (no jobs array)
-  const badResponses = [
-    [{}, 'empty object'],
-    [{ jobs: null }, 'jobs: null'],
-    [{ jobs: 'not-array' }, 'jobs: string'],
-    [{ offers: [] }, 'wrong key name'],
-    [null, 'null response'],
-  ];
-  for (const [resp, label] of badResponses) {
-    let threw = false;
-    try {
-      await sj.fetch(
-        { name: 'SolidJobs IT', careers_url: 'https://solid.jobs/public-api/offers/it?campaign=career-ops' },
-        { transport: 'http', fetchJson: async () => resp, fetchText: async () => '' },
-      );
-    } catch (e) {
-      if (e.message.includes('unexpected API response')) threw = true;
-      else fail(`solidjobs.fetch() bad response (${label}) wrong error: ${e.message}`);
-    }
-    if (threw) pass(`solidjobs.fetch() throws on bad API response (${label})`);
-    else fail(`solidjobs.fetch() should throw on bad API response (${label})`);
-  }
-
-  // fetch() filters out jobs with empty/missing url
-  const mixedJobs = {
-    jobs: [
-      { title: 'Has URL', url: 'https://solid.jobs/o/1/career-ops', company: 'A', locations: [] },
-      { title: 'No URL', url: '', company: 'B', locations: [] },
-      { title: 'Missing URL', company: 'C', locations: [] },
-    ],
-  };
-  const filtered = await sj.fetch(
-    { name: 'SolidJobs IT', careers_url: 'https://solid.jobs/public-api/offers/it?campaign=career-ops' },
-    { transport: 'http', fetchJson: async () => mixedJobs, fetchText: async () => '' },
-  );
-  if (filtered.length === 1 && filtered[0].title === 'Has URL') pass('solidjobs.fetch() filters out jobs with empty/missing url');
-  else fail(`solidjobs.fetch() should filter empty URLs, got ${filtered.length} jobs: ${JSON.stringify(filtered)}`);
-
-  // fetch() handles string locations (non-array)
-  const stringLocJobs = { jobs: [{ title: 'Dev', url: 'https://solid.jobs/o/2/career-ops', company: 'X', locations: 'Warsaw' }] };
-  const strLoc = await sj.fetch(
-    { name: 'SolidJobs IT', careers_url: 'https://solid.jobs/public-api/offers/it?campaign=career-ops' },
-    { transport: 'http', fetchJson: async () => stringLocJobs, fetchText: async () => '' },
-  );
-  if (strLoc[0].location === 'Warsaw') pass('solidjobs.fetch() handles string locations');
-  else fail(`solidjobs.fetch() string location is ${JSON.stringify(strLoc[0].location)}, expected "Warsaw"`);
-
-  // detect() returns null for valid hostname but wrong path
-  if (sj.detect({ name: 'X', careers_url: 'https://solid.jobs/careers' }) === null) {
-    pass('solidjobs.detect() rejects solid.jobs URL with wrong path');
-  } else {
-    fail('solidjobs.detect() should reject solid.jobs URLs not under /public-api/offers/');
-  }
-
-  // fetch() passes redirect:'error' to fetchJson
-  let capturedOpts = null;
-  await sj.fetch(
-    { name: 'SolidJobs IT', careers_url: 'https://solid.jobs/public-api/offers/it?campaign=career-ops' },
-    { transport: 'http', fetchJson: async (_url, opts) => { capturedOpts = opts; return { jobs: [] }; }, fetchText: async () => '' },
-  );
-  if (capturedOpts && capturedOpts.redirect === 'error') pass('solidjobs.fetch() passes redirect:"error" to fetchJson');
-  else fail(`solidjobs.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-
-  // fetch() tolerates malformed array members without crashing
-  const malformedMembers = { jobs: [null, 7, { title: 'OK', url: 'https://solid.jobs/o/3/career-ops', company: 'Z' }] };
-  const safeParsed = await sj.fetch(
-    { name: 'SolidJobs IT', careers_url: 'https://solid.jobs/public-api/offers/it?campaign=career-ops' },
-    { transport: 'http', fetchJson: async () => malformedMembers, fetchText: async () => '' },
-  );
-  if (safeParsed.length === 1 && safeParsed[0].url === 'https://solid.jobs/o/3/career-ops') {
-    pass('solidjobs.fetch() skips malformed jobs members without crashing');
-  } else {
-    fail(`solidjobs.fetch() malformed members handling failed: ${JSON.stringify(safeParsed)}`);
-  }
-} catch (e) {
-  fail(`solidjobs provider tests crashed: ${e.message}`);
-}
-
-// ── 16. SSRF redirect hardening (lever / ashby) ──────────────────
-// _http.mjs defaults to redirect:'follow', so a server-side redirect from any
-// of these ATS APIs to an internal address is an SSRF vector. Every other GET
-// provider passes redirect:'error'; these two were missing it.
-// (workday's redirect:'error' coverage lives in its own "Provider — workday"
-// section below, checked across every paginated request, not just the first.)
-
-console.log('\n16. Provider — SSRF redirect hardening (lever / ashby)');
-
-try {
-  const lever = (await import(pathToFileURL(join(ROOT, 'providers/lever.mjs')).href)).default;
-  const ashby = (await import(pathToFileURL(join(ROOT, 'providers/ashby.mjs')).href)).default;
-
-  let leverOpts = null;
-  await lever.fetch(
-    { name: 'L', careers_url: 'https://jobs.lever.co/example' },
-    { transport: 'http', fetchJson: async (_u, opts) => { leverOpts = opts; return []; }, fetchText: async () => '' },
-  );
-  if (leverOpts && leverOpts.redirect === 'error') pass('lever.fetch() passes redirect:"error"');
-  else fail(`lever.fetch() should pass redirect:"error", got ${JSON.stringify(leverOpts)}`);
-
-  let ashbyOpts = null;
-  await ashby.fetch(
-    { name: 'A', careers_url: 'https://jobs.ashbyhq.com/example' },
-    { transport: 'http', fetchJson: async (_u, opts) => { ashbyOpts = opts; return { jobs: [] }; }, fetchText: async () => '' },
-  );
-  if (ashbyOpts && ashbyOpts.redirect === 'error') pass('ashby.fetch() passes redirect:"error"');
-  else fail(`ashby.fetch() should pass redirect:"error", got ${JSON.stringify(ashbyOpts)}`);
-} catch (e) {
-  fail(`SSRF redirect hardening tests crashed: ${e.message}`);
-}
-
 // ── 15. URL REDISCOVERY FALLBACK (--rediscover-404) ─────────────
 
 console.log('\n15. URL rediscovery fallback');
@@ -5518,594 +4788,6 @@ try {
   }
 } catch (e) {
   fail(`CJK rendering test crashed: ${e.message}`);
-}
-
-// ── 22. PROVIDERS — Jobstreet (SEEK v5 JobSearch API) ────────────
-
-console.log('\n22. Provider — jobstreet');
-
-try {
-  const jobstreet = (await import(pathToFileURL(join(ROOT, 'providers/jobstreet.mjs')).href)).default;
-  const { parseJobstreetItem } = await import(pathToFileURL(join(ROOT, 'providers/jobstreet.mjs')).href);
-
-  // id check
-  if (jobstreet.id === 'jobstreet') pass('jobstreet.id is "jobstreet"');
-  else fail(`jobstreet.id is ${JSON.stringify(jobstreet.id)}`);
-
-  // detect() always returns null (job board, not ATS)
-  if (jobstreet.detect({ name: 'X', careers_url: 'https://id.jobstreet.com/jobs' }) === null) {
-    pass('jobstreet.detect() returns null — explicit provider only, no URL auto-detection');
-  } else {
-    fail('jobstreet.detect() should return null for any URL');
-  }
-
-  // parseJobstreetItem — valid item (v5 API shape)
-  const sampleItem = {
-    id: '92996157',
-    title: 'Facility Engineer',
-    advertiser: { id: '60960115', description: 'PT YOFC International Indonesia' },
-    companyName: 'YOFC International',
-    locations: [{ label: 'Karawang, West Java', countryCode: 'ID' }],
-    listingDate: '2026-06-29T02:53:00Z',
-  };
-  const parsed = parseJobstreetItem(sampleItem, 'https://id.jobstreet.com', 'FallbackCo');
-  if (parsed && parsed.title === 'Facility Engineer'
-      && parsed.url === 'https://id.jobstreet.com/id/job/92996157'
-      && parsed.company === 'PT YOFC International Indonesia'
-      && parsed.location === 'Karawang, West Java'
-      && parsed.postedAt != null) {
-    pass('parseJobstreetItem extracts title, url, company, location, postedAt correctly');
-  } else {
-    fail(`parseJobstreetItem returned ${JSON.stringify(parsed)}`);
-  }
-
-  // parseJobstreetItem — uses companyName fallback when advertiser.description is absent
-  const noAdvertiserItem = {
-    id: '2',
-    title: 'Data Scientist',
-    companyName: 'TechCorp',
-    locations: [{ label: 'Jakarta' }],
-    listingDate: '2026-06-01T00:00:00Z',
-  };
-  const noAdvParsed = parseJobstreetItem(noAdvertiserItem, 'https://id.jobstreet.com', 'Fallback');
-  if (noAdvParsed && noAdvParsed.company === 'TechCorp') {
-    pass('parseJobstreetItem falls back to companyName when advertiser.description is absent');
-  } else {
-    fail(`parseJobstreetItem companyName fallback: ${JSON.stringify(noAdvParsed)}`);
-  }
-
-  // parseJobstreetItem — rejects items without title
-  if (parseJobstreetItem({ id: '1' }, 'https://id.jobstreet.com', 'Co') === null) {
-    pass('parseJobstreetItem returns null for items without title');
-  } else {
-    fail('parseJobstreetItem should return null for title-less items');
-  }
-
-  // parseJobstreetItem — rejects items without id (URL cannot be built)
-  if (parseJobstreetItem({ title: 'Role' }, 'https://id.jobstreet.com', 'Co') === null) {
-    pass('parseJobstreetItem returns null for items without id');
-  } else {
-    fail('parseJobstreetItem should return null for items without id');
-  }
-
-  // parseJobstreetItem — rejects off-domain base URLs
-  const offDomain = parseJobstreetItem(
-    { id: '1', title: 'Role', locations: [{ label: 'Remote' }] },
-    'https://evil.example.com', 'Co'
-  );
-  if (offDomain === null) pass('parseJobstreetItem rejects off-domain base URLs');
-  else fail(`parseJobstreetItem should reject off-domain base URLs, got ${JSON.stringify(offDomain)}`);
-
-  // parseJobstreetItem — handles null/malformed input safely
-  if (parseJobstreetItem(null, 'https://id.jobstreet.com', 'Co') === null) pass('parseJobstreetItem(null) → null');
-  else fail('parseJobstreetItem(null) should return null');
-  if (parseJobstreetItem(7, 'https://id.jobstreet.com', 'Co') === null) pass('parseJobstreetItem(7) → null');
-  else fail('parseJobstreetItem(number) should return null');
-
-  // parseJobstreetItem — uses fallback company when both advertiser and companyName are missing
-  const noCompany = parseJobstreetItem(
-    { id: '99', title: 'Engineer', locations: [{ label: 'Remote' }] },
-    'https://id.jobstreet.com', 'PortalFallback'
-  );
-  if (noCompany && noCompany.company === 'PortalFallback') {
-    pass('parseJobstreetItem uses fallback company when all company fields are missing');
-  } else {
-    fail(`parseJobstreetItem fallback company: ${JSON.stringify(noCompany)}`);
-  }
-
-  // parseJobstreetItem — handles empty locations gracefully
-  const noLocItem = {
-    id: '42',
-    title: 'Remote Engineer',
-    advertiser: { description: 'RemoteCo' },
-    listingDate: '2026-06-15T00:00:00Z',
-  };
-  const noLocParsed = parseJobstreetItem(noLocItem, 'https://id.jobstreet.com', 'Fallback');
-  if (noLocParsed && noLocParsed.location === '') {
-    pass('parseJobstreetItem handles missing locations gracefully');
-  } else {
-    fail(`parseJobstreetItem empty location: ${JSON.stringify(noLocParsed)}`);
-  }
-
-  // fetch() — happy path with mock context (v5 API response shape)
-  const mockCtx = {
-    transport: 'http',
-    fetchJson: async (url) => {
-      if (!url.startsWith('https://id.jobstreet.com/api/jobsearch/v5/search')) throw new Error('Unexpected URL');
-      return {
-        data: [
-          {
-            id: '92884969',
-            title: 'AI Engineer',
-            advertiser: { id: '14025083', description: 'Capgemini' },
-            companyName: 'Capgemini',
-            locations: [{ label: 'Jakarta, Indonesia', countryCode: 'ID' }],
-            listingDate: '2026-06-23T03:55:20Z',
-          },
-        ],
-        totalCount: 1,
-      };
-    },
-    fetchText: async () => { throw new Error('should not be called'); },
-  };
-  const jobs = await jobstreet.fetch(
-    { name: 'Jobstreet ID', provider: 'jobstreet', searchKeywords: 'AI' },
-    mockCtx,
-  );
-  if (jobs.length === 1 && jobs[0].title === 'AI Engineer') pass('jobstreet.fetch() returns parsed jobs via v5 API');
-  else fail(`jobstreet.fetch() returned ${JSON.stringify(jobs)}`);
-
-  // fetch() — handles empty results
-  const emptyCtx = {
-    transport: 'http',
-    fetchJson: async () => ({ data: [], totalCount: 0 }),
-    fetchText: async () => { throw new Error('should not be called'); },
-  };
-  const emptyJobs = await jobstreet.fetch(
-    { name: 'Jobstreet ID', provider: 'jobstreet', searchKeywords: 'nonexistent' },
-    emptyCtx,
-  );
-  if (emptyJobs.length === 0) pass('jobstreet.fetch() handles empty results');
-  else fail(`jobstreet.fetch() should return empty array for no results, got ${emptyJobs.length}`);
-
-  // fetch() — rejects invalid hostname
-  let hostRejected = false;
-  try {
-    await jobstreet.fetch(
-      { name: 'Bad', provider: 'jobstreet', api: 'https://evil.example.com/api/jobsearch/v5/search' },
-      { transport: 'http', fetchJson: async () => ({}), fetchText: async () => '' },
-    );
-  } catch (e) {
-    if (e.message.includes('untrusted hostname')) hostRejected = true;
-    else fail(`jobstreet.fetch() host rejection wrong error: ${e.message}`);
-  }
-  if (hostRejected) pass('jobstreet.fetch() rejects untrusted hostnames');
-  else fail('jobstreet.fetch() should reject non-jobstreet hostnames');
-
-  // fetch() — handles non-array data field
-  const badDataCtx = {
-    transport: 'http',
-    fetchJson: async () => ({ data: null }),
-    fetchText: async () => { throw new Error('should not be called'); },
-  };
-  const badDataJobs = await jobstreet.fetch(
-    { name: 'Jobstreet ID', provider: 'jobstreet', searchKeywords: 'test' },
-    badDataCtx,
-  );
-  if (badDataJobs.length === 0) pass('jobstreet.fetch() handles null data field');
-  else fail(`jobstreet.fetch() should return empty for null data`);
-
-} catch (e) {
-  fail(`jobstreet provider tests crashed: ${e.message}`);
-}
-
-// ── 23. PROVIDERS — Glints (v2-alc / searchJobsV3) ────────────────
-
-console.log('\n23. Provider — glints');
-
-try {
-  const glints = (await import(pathToFileURL(join(ROOT, 'providers/glints.mjs')).href)).default;
-  const { parseGlintsItem } = await import(pathToFileURL(join(ROOT, 'providers/glints.mjs')).href);
-
-  // id check
-  if (glints.id === 'glints') pass('glints.id is "glints"');
-  else fail(`glints.id is ${JSON.stringify(glints.id)}`);
-
-  // detect() always returns null (job board, not ATS)
-  if (glints.detect({ name: 'X', careers_url: 'https://glints.com/id/jobs' }) === null) {
-    pass('glints.detect() returns null — explicit provider only, no URL auto-detection');
-  } else {
-    fail('glints.detect() should return null for any URL');
-  }
-
-  // parseGlintsItem — valid item (new searchJobsV3 shape)
-  const sampleItem = {
-    id: 'abc123',
-    title: 'Backend Engineer',
-    company: { name: 'StartupCorp', brandName: 'StartupCorp Brand' },
-    city: { name: 'Jakarta, Indonesia' },
-    country: { code: 'ID', name: 'Indonesia' },
-    createdAt: '2026-06-10T00:00:00Z',
-  };
-  const parsed = parseGlintsItem(sampleItem, 'https://glints.com', 'FallbackCo');
-  if (parsed && parsed.title === 'Backend Engineer'
-      && parsed.url === 'https://glints.com/id/opportunities/jobs/abc123'
-      && parsed.company === 'StartupCorp'
-      && parsed.location === 'Jakarta, Indonesia'
-      && parsed.postedAt != null) {
-    pass('parseGlintsItem extracts title, url, company, location, postedAt correctly');
-  } else {
-    fail(`parseGlintsItem returned ${JSON.stringify(parsed)}`);
-  }
-
-  // parseGlintsItem — uses brandName when name is absent
-  const brandItem = {
-    id: 'def456',
-    title: 'Data Scientist',
-    company: { brandName: 'BrandCorp' },
-    city: { name: 'Singapore' },
-    createdAt: '2026-06-15T00:00:00Z',
-  };
-  const brandParsed = parseGlintsItem(brandItem, 'https://glints.com', 'FallbackCo');
-  if (brandParsed && brandParsed.company === 'BrandCorp') {
-    pass('parseGlintsItem falls back to brandName when company.name is missing');
-  } else {
-    fail(`parseGlintsItem brandName fallback: ${JSON.stringify(brandParsed)}`);
-  }
-
-  // parseGlintsItem — rejects items without title
-  if (parseGlintsItem({ id: '1' }, 'https://glints.com', 'Co') === null) {
-    pass('parseGlintsItem returns null for title-less items');
-  } else {
-    fail('parseGlintsItem should return null for items without title');
-  }
-
-  // parseGlintsItem — rejects items without id (URL cannot be built)
-  if (parseGlintsItem({ title: 'Role' }, 'https://glints.com', 'Co') === null) {
-    pass('parseGlintsItem returns null for items without id');
-  } else {
-    fail('parseGlintsItem should return null for items without id');
-  }
-
-  // parseGlintsItem — rejects off-domain via URL validation
-  const offDomain = parseGlintsItem(
-    { id: '1', title: 'Role' },
-    'https://evil.example.com', 'Co'
-  );
-  if (offDomain === null) pass('parseGlintsItem rejects off-domain base URLs');
-  else fail(`parseGlintsItem should reject off-domain base URLs, got ${JSON.stringify(offDomain)}`);
-
-  // parseGlintsItem — handles null/malformed input
-  if (parseGlintsItem(null, 'https://glints.com', 'Co') === null) pass('parseGlintsItem(null) → null');
-  else fail('parseGlintsItem(null) should return null');
-  if (parseGlintsItem(42, 'https://glints.com', 'Co') === null) pass('parseGlintsItem(number) → null');
-  else fail('parseGlintsItem(number) should return null');
-
-  // parseGlintsItem — fallback company when both company.name and brandName are missing
-  const noCompany = parseGlintsItem(
-    { id: '99', title: 'Engineer' },
-    'https://glints.com', 'PortalName'
-  );
-  if (noCompany && noCompany.company === 'PortalName') {
-    pass('parseGlintsItem uses fallback company when company info is absent');
-  } else {
-    fail(`parseGlintsItem fallback company: ${JSON.stringify(noCompany)}`);
-  }
-
-  // parseGlintsItem — parseGlintsItem allows www.glints.com subdomain via baseUrl
-  const wwwItem = parseGlintsItem(
-    { id: 'x', title: 'Role' },
-    'https://www.glints.com', 'Co'
-  );
-  if (wwwItem && wwwItem.url.startsWith('https://www.glints.com/')) {
-    pass('parseGlintsItem accepts www.glints.com base URL (subdomain)');
-  } else {
-    fail(`parseGlintsItem www subdomain: ${JSON.stringify(wwwItem)}`);
-  }
-
-  // fetch() — happy path with mock context (searchJobsV3 response shape)
-  const mockCtx = {
-    transport: 'http',
-    fetchJson: async (url, opts) => {
-      if (opts?.method !== 'POST') throw new Error('Expected POST');
-      const body = JSON.parse(opts.body || '{}');
-      if (!body.query) throw new Error('Expected GraphQL query');
-      if (body.operationName !== 'searchJobsV3') throw new Error('Expected searchJobsV3 operation');
-      return {
-        data: {
-          searchJobsV3: {
-            jobsInPage: [
-              {
-                id: 'job1',
-                title: 'AI PM',
-                company: { name: 'TechCo', brandName: 'TechCo Brand' },
-                city: { name: 'Remote' },
-                country: { code: 'ID', name: 'Indonesia' },
-                salaries: [],
-                createdAt: '2026-01-01T00:00:00Z',
-              },
-            ],
-            expInfo: null,
-            hasMore: false,
-          },
-        },
-      };
-    },
-    fetchText: async () => { throw new Error('should not be called'); },
-  };
-  const jobs = await glints.fetch(
-    { name: 'Glints ID', provider: 'glints', searchKeywords: 'AI' },
-    mockCtx,
-  );
-  if (jobs.length === 1 && jobs[0].title === 'AI PM') pass('glints.fetch() returns parsed jobs via searchJobsV3');
-  else fail(`glints.fetch() returned ${JSON.stringify(jobs)}`);
-
-  // fetch() — handles empty results
-  const emptyCtx = {
-    transport: 'http',
-    fetchJson: async () => ({ data: { searchJobsV3: { jobsInPage: [], hasMore: false } } }),
-    fetchText: async () => { throw new Error('should not be called'); },
-  };
-  const emptyJobs = await glints.fetch(
-    { name: 'Glints ID', provider: 'glints', searchKeywords: 'nonexistent' },
-    emptyCtx,
-  );
-  if (emptyJobs.length === 0) pass('glints.fetch() handles empty results');
-  else fail(`glints.fetch() should return empty array for no results, got ${emptyJobs.length}`);
-
-  // fetch() — stops when hasMore is false (single page)
-  const singlePageCtx = {
-    transport: 'http',
-    fetchJson: async () => ({ data: { searchJobsV3: { jobsInPage: [{ id: '1', title: 'Dev' }], hasMore: false } } }),
-    fetchText: async () => { throw new Error('should not be called'); },
-  };
-  const singleJobs = await glints.fetch(
-    { name: 'Glints ID', provider: 'glints', searchKeywords: 'dev' },
-    singlePageCtx,
-  );
-  if (singleJobs.length === 1) pass('glints.fetch() stops when hasMore is false');
-  else fail(`glints.fetch() single page: ${JSON.stringify(singleJobs)}`);
-
-  // fetch() — rejects invalid hostname
-  let hostRejected = false;
-  try {
-    await glints.fetch(
-      { name: 'Bad', provider: 'glints', api: 'https://evil.example.com/graphql' },
-      { transport: 'http', fetchJson: async () => ({}), fetchText: async () => '' },
-    );
-  } catch (e) {
-    if (e.message.includes('untrusted hostname')) hostRejected = true;
-    else fail(`glints.fetch() host rejection wrong error: ${e.message}`);
-  }
-  if (hostRejected) pass('glints.fetch() rejects untrusted hostnames');
-  else fail('glints.fetch() should reject non-glints hostnames');
-
-  // fetch() — throws on missing searchJobsV3 in response
-  let missingThrew = false;
-  try {
-    await glints.fetch(
-      { name: 'Glints ID', provider: 'glints', searchKeywords: 'test' },
-      {
-        transport: 'http',
-        fetchJson: async () => ({ data: { somethingElse: [] } }),
-        fetchText: async () => { throw new Error('should not be called'); },
-      },
-    );
-  } catch (e) {
-    if (e.message.includes('unexpected API response')) missingThrew = true;
-    else fail(`glints.fetch() missing searchJobsV3 wrong error: ${e.message}`);
-  }
-  if (missingThrew) pass('glints.fetch() throws on unexpected API response shape');
-  else fail('glints.fetch() should throw when searchJobsV3 is missing');
-
-} catch (e) {
-  fail(`glints provider tests crashed: ${e.message}`);
-}
-
-console.log('\n25. Provider — arbeitsagentur');
-
-try {
-  const aa = (await import(pathToFileURL(join(ROOT, 'providers/arbeitsagentur.mjs')).href)).default;
-  const { parseArbeitsagenturConfig, buildLocation, normalizeJob } =
-    await import(pathToFileURL(join(ROOT, 'providers/arbeitsagentur.mjs')).href);
-
-  if (aa.id === 'arbeitsagentur') pass('arbeitsagentur.id is "arbeitsagentur"');
-  else fail(`arbeitsagentur.id is ${JSON.stringify(aa.id)}`);
-
-  // parseArbeitsagenturConfig — defaults when block is absent
-  const def = parseArbeitsagenturConfig({});
-  if (def.keywords.length === 0 && def.wo === '' && def.umkreis === 50 && def.days === 30 && def.size === 100 && def.remoteNationwide === false) {
-    pass('parseArbeitsagenturConfig applies defaults (umkreis 50, days 30, size 100)');
-  } else {
-    fail(`parseArbeitsagenturConfig defaults = ${JSON.stringify(def)}`);
-  }
-
-  // parseArbeitsagenturConfig — sanitizes keywords and clamps numbers
-  const cfg = parseArbeitsagenturConfig({
-    arbeitsagentur: { keywords: ['  ML Engineer  ', '', 7, 'NLP'], wo: ' Berlin ', umkreis: 999999, size: 0, days: -3, remoteNationwide: 'yes' },
-  });
-  if (cfg.keywords.length === 2 && cfg.keywords[0] === 'ML Engineer' && cfg.keywords[1] === 'NLP') {
-    pass('parseArbeitsagenturConfig trims keywords and drops empty/non-string entries');
-  } else {
-    fail(`parseArbeitsagenturConfig keywords = ${JSON.stringify(cfg.keywords)}`);
-  }
-  if (cfg.wo === 'Berlin' && cfg.umkreis === 1000 && cfg.size === 1 && cfg.days === 1 && cfg.remoteNationwide === false) {
-    pass('parseArbeitsagenturConfig clamps umkreis/size/days and treats non-true remoteNationwide as false');
-  } else {
-    fail(`parseArbeitsagenturConfig sanitized = ${JSON.stringify(cfg)}`);
-  }
-
-  // buildLocation — ort/region join, non-DE country appended, DE omitted
-  if (buildLocation({ ort: 'Berlin', region: 'Berlin', land: 'Deutschland' }) === 'Berlin, Berlin') {
-    pass('buildLocation joins ort/region and omits Germany');
-  } else {
-    fail(`buildLocation DE = ${JSON.stringify(buildLocation({ ort: 'Berlin', region: 'Berlin', land: 'Deutschland' }))}`);
-  }
-  if (buildLocation({ ort: 'Wien', land: 'Österreich' }) === 'Wien, Österreich') {
-    pass('buildLocation appends non-DE country');
-  } else {
-    fail(`buildLocation non-DE = ${JSON.stringify(buildLocation({ ort: 'Wien', land: 'Österreich' }))}`);
-  }
-  if (buildLocation(null) === '' && buildLocation('x') === '') pass('buildLocation returns "" for missing/garbage input');
-  else fail('buildLocation should return "" for missing/garbage input');
-
-  // normalizeJob — happy path encodes refnr into the detail URL
-  const norm = normalizeJob({ refnr: '10000-123/4 X', titel: '  ML Engineer  ', arbeitgeber: ' ACME ', arbeitsort: { ort: 'Berlin' } });
-  if (norm && norm.title === 'ML Engineer' && norm.company === 'ACME'
-      && norm.url === 'https://www.arbeitsagentur.de/jobsuche/jobdetail/' + encodeURIComponent('10000-123/4 X')
-      && norm.refnr === '10000-123/4 X') {
-    pass('normalizeJob trims fields and URL-encodes refnr');
-  } else {
-    fail(`normalizeJob = ${JSON.stringify(norm)}`);
-  }
-  if (normalizeJob({ titel: 'No refnr' }) === null && normalizeJob({ refnr: 'x', titel: '' }) === null) {
-    pass('normalizeJob returns null without a refnr or title');
-  } else {
-    fail('normalizeJob should return null when refnr or title is missing');
-  }
-
-  // fetch() — nationwide single-keyword pass, dedup across keywords, header sent
-  let sentApiKey = null;
-  const mkCtx = (byWas) => ({
-    fetchJson: async (url, opts) => {
-      sentApiKey = opts?.headers?.['X-API-Key'] ?? sentApiKey;
-      const was = new URL(url).searchParams.get('was');
-      return { stellenangebote: byWas[was] || [] };
-    },
-  });
-  const fetched = await aa.fetch(
-    { name: 'AA', arbeitsagentur: { keywords: ['ML', 'NLP'] } },
-    mkCtx({
-      ML: [{ refnr: 'A', titel: 'ML Engineer', arbeitgeber: 'Co', arbeitsort: { ort: 'Berlin' } }],
-      NLP: [
-        { refnr: 'A', titel: 'ML Engineer', arbeitgeber: 'Co', arbeitsort: { ort: 'Berlin' } }, // dup refnr
-        { refnr: 'B', titel: 'NLP Scientist', arbeitgeber: 'Co', arbeitsort: { ort: 'Köln' } },
-      ],
-    }),
-  );
-  if (fetched.length === 2 && !('refnr' in fetched[0])) pass('aa.fetch() dedups by refnr and strips refnr from output');
-  else fail(`aa.fetch() returned ${JSON.stringify(fetched)}`);
-  if (sentApiKey === 'jobboerse-jobsuche') pass('aa.fetch() sends the X-API-Key header');
-  else fail(`aa.fetch() X-API-Key = ${JSON.stringify(sentApiKey)}`);
-
-  // fetch() — remoteNationwide pass keeps only remote-titled wide hits
-  let calls = 0;
-  const remoteFetched = await aa.fetch(
-    { name: 'AA', arbeitsagentur: { keywords: ['ML'], wo: 'Berlin', remoteNationwide: true } },
-    {
-      fetchJson: async (url) => {
-        calls++;
-        const hasWo = new URL(url).searchParams.has('wo');
-        // Pass A (wo set) → local hit; Pass B (no wo) → one remote-titled, one not.
-        return hasWo
-          ? { stellenangebote: [{ refnr: 'L', titel: 'ML Engineer', arbeitgeber: 'Co', arbeitsort: { ort: 'Berlin' } }] }
-          : { stellenangebote: [
-              { refnr: 'R', titel: 'ML Engineer (Remote)', arbeitgeber: 'Co', arbeitsort: { ort: 'Hamburg' } },
-              { refnr: 'X', titel: 'Onsite ML Engineer', arbeitgeber: 'Co', arbeitsort: { ort: 'Hamburg' } },
-            ] };
-      },
-    },
-  );
-  if (calls === 2 && remoteFetched.some(j => j.url.endsWith('R')) && !remoteFetched.some(j => j.url.endsWith('X'))) {
-    pass('aa.fetch() remoteNationwide keeps remote-titled wide hits and drops onsite ones');
-  } else {
-    fail(`aa.fetch() remoteNationwide = ${calls} calls, ${JSON.stringify(remoteFetched.map(j => j.url))}`);
-  }
-
-  // parseArbeitsagenturConfig — remoteMatch mode + remoteMaxPages (config-driven remote detection)
-  const rcfg = parseArbeitsagenturConfig({ arbeitsagentur: { keywords: ['ML'], remoteMatch: 'filter', remoteMaxPages: 50 } });
-  if (rcfg.remoteMatch === 'filter' && rcfg.remoteMaxPages === 20) {
-    pass('parseArbeitsagenturConfig parses remoteMatch and clamps remoteMaxPages');
-  } else {
-    fail(`parseArbeitsagenturConfig remoteMatch/maxPages = ${JSON.stringify({ m: rcfg.remoteMatch, p: rcfg.remoteMaxPages })}`);
-  }
-  const rdef = parseArbeitsagenturConfig({ arbeitsagentur: { keywords: ['ML'], remoteMatch: 'bogus' } });
-  if (rdef.remoteMatch === 'title' && rdef.remoteMaxPages === 1) {
-    pass('parseArbeitsagenturConfig defaults remoteMatch to "title" and remoteMaxPages to 1');
-  } else {
-    fail(`parseArbeitsagenturConfig remote defaults = ${JSON.stringify({ m: rdef.remoteMatch, p: rdef.remoteMaxPages })}`);
-  }
-
-  // fetch() — remoteMatch:'filter' uses server-side homeoffice filter, paginates, and tags remote roles
-  let usedHomeoffice = false;
-  const pagesSeen = new Set();
-  const filterFetched = await aa.fetch(
-    { name: 'AA', arbeitsagentur: { keywords: ['ML'], wo: 'Berlin', remoteNationwide: true, remoteMatch: 'filter', remoteMaxPages: 5, size: 2 } },
-    {
-      fetchJson: async (url) => {
-        const sp = new URL(url).searchParams;
-        if (sp.has('wo')) {
-          return { stellenangebote: [{ refnr: 'L', titel: 'ML Engineer', arbeitgeber: 'Co', arbeitsort: { ort: 'Berlin' } }] };
-        }
-        usedHomeoffice = usedHomeoffice || sp.get('homeoffice') === 'nv_true';
-        pagesSeen.add(sp.get('page'));
-        return Number(sp.get('page')) === 1
-          ? { stellenangebote: [ // full page (== size) → pagination continues
-              { refnr: 'R1', titel: 'ML Engineer', arbeitgeber: 'Co', arbeitsort: { ort: 'München' } },
-              { refnr: 'R2', titel: 'ML Scientist', arbeitgeber: 'Co', arbeitsort: { ort: 'Stuttgart' } },
-            ] }
-          : { stellenangebote: [{ refnr: 'R3', titel: 'NLP Engineer', arbeitgeber: 'Co', arbeitsort: { ort: 'Köln' } }] }; // short → stop
-      },
-    },
-  );
-  const munich = filterFetched.find(j => j.url.endsWith('R1'));
-  if (usedHomeoffice && pagesSeen.has('1') && pagesSeen.has('2') && munich && /Deutschlandweit \(Homeoffice\)/.test(munich.location)) {
-    pass('aa.fetch() remoteMatch:filter sends homeoffice=nv_true, paginates, and tags far-city remote roles');
-  } else {
-    fail(`aa.fetch() filter mode = ${JSON.stringify({ usedHomeoffice, pages: [...pagesSeen], munichLoc: munich?.location })}`);
-  }
-
-  // fetch() — no keywords throws; total outage throws (not silent)
-  let noKw = false;
-  try { await aa.fetch({ name: 'AA', arbeitsagentur: {} }, mkCtx({})); } catch { noKw = true; }
-  if (noKw) pass('aa.fetch() throws when no keywords are configured');
-  else fail('aa.fetch() should throw without keywords');
-
-  let outage = false;
-  try {
-    await aa.fetch({ name: 'AA', arbeitsagentur: { keywords: ['ML'] } }, { fetchJson: async () => { throw new Error('HTTP 503'); } });
-  } catch { outage = true; }
-  if (outage) pass('aa.fetch() throws when every keyword request fails (no silent empty)');
-  else fail('aa.fetch() should throw on total outage');
-
-  // fetch() — one keyword answers (empty) while another fails → NOT a total
-  // outage; partial success must not throw.
-  let partialThrew = false;
-  let partial;
-  try {
-    partial = await aa.fetch(
-      { name: 'AA', arbeitsagentur: { keywords: ['OK', 'BAD'] } },
-      { fetchJson: async (url) => {
-          if (new URL(url).searchParams.get('was') === 'BAD') throw new Error('HTTP 503');
-          return { stellenangebote: [] }; // OK answers, just empty
-        } },
-    );
-  } catch { partialThrew = true; }
-  if (!partialThrew && Array.isArray(partial) && partial.length === 0) {
-    pass('aa.fetch() does not throw when one keyword succeeds empty and another fails');
-  } else {
-    fail(`aa.fetch() partial-success threw=${partialThrew}, result=${JSON.stringify(partial)}`);
-  }
-
-  // fetch() — Pass A succeeds with jobs, optional Pass B fails → Pass A jobs kept.
-  const passBFail = await aa.fetch(
-    { name: 'AA', arbeitsagentur: { keywords: ['ML'], wo: 'Berlin', remoteNationwide: true } },
-    { fetchJson: async (url) => {
-        // Pass A (wo set) returns a job; Pass B (no wo) throws.
-        if (new URL(url).searchParams.has('wo')) {
-          return { stellenangebote: [{ refnr: 'L', titel: 'ML Engineer', arbeitgeber: 'Co', arbeitsort: { ort: 'Berlin' } }] };
-        }
-        throw new Error('HTTP 503');
-      } },
-  );
-  if (passBFail.length === 1 && passBFail[0].url.endsWith('L')) {
-    pass('aa.fetch() preserves primary (Pass A) results when the remote pass (Pass B) fails');
-  } else {
-    fail(`aa.fetch() Pass B failure dropped primary: ${JSON.stringify(passBFail)}`);
-  }
-
-} catch (e) {
-  fail(`arbeitsagentur provider tests crashed: ${e.message}`);
 }
 
 console.log('\n24. Provider — ibm');
